計算精度を上げるため変数型を変更

+++ CalorieCalculator/AmountOfIntake.cs
@@ -8,16 +8,16 @@
         /// <summary>
         /// タンパク質
         /// </summary>
-        public double Protein { get; init; }
+        public decimal Protein { get; init; }
 
         /// <summary>
         /// 脂肪
         /// </summary>
-        public double Fat { get; init; }
+        public decimal Fat { get; init; }
 
         /// <summary>
         /// 炭水化物
         /// </summary>
-        public double Carbohydrate { get; init; }
+        public decimal Carbohydrate { get; init; }
     }
 }
