+++ CalorieCalculator/ICalculateCarbohydrateCalories.cs
@@ -2,6 +2,6 @@
 {
     public interface ICalculateCarbohydrateCalories
     {
-        long Execute(double carbohydrateIntake);
+        long Execute(decimal carbohydrateIntake);
     }
 }
